requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
google-api-python-client==2.108.0
torch==2.1.0
transformers==4.35.0
scipy==1.11.4
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
python-dotenv==1.0.0
httpx==0.25.0